\name{lamle-package}
\alias{lamle-package}
\docType{package}
\title{
\packageTitle{lamle}
}
\description{
\packageDescription{lamle}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{lamle}
\packageIndices{lamle}

}
\author{
Björn Andersson, Shaobo Jin and Maoxin Zhang.

Maintainer: Björn Andersson <bjoern.h.andersson@gmail.com>
}
\references{
Andersson, B., Jin, S., and Zhang, M. (2023). Fast estimation of multiple group generalized linear latent variable models for categorical observed variables. \emph{Computational Statistics and Data Analysis}, \bold{182}, 1-12. <doi:10.1016/j.csda.2023.107710>

Andersson, B., and Xin, T. (2021). Estimation of latent regression item response theory models using a second-order Laplace approximation. \emph{Journal of Educational and Behavioral Statistics}, \bold{46(2)}, 244-265. <doi:10.3102/1076998620945199>

Huber, P., Ronchetti, E., and Victoria-Feser, M.P. (2004). Estimation of generalized linear latent variable models. \emph{Journal of the Royal Statistical Society: Series B (Statistical Methodology)}, \bold{66(4)}, 893-908. <doi:10.1111/j.1467-9868.2004.05627.x>

Jin, S., Noh, M., and Lee, Y. (2018). H-Likelihood Approach to Factor Analysis for Ordinal Data. \emph{Structural Equation Modeling: A Multidisciplinary Journal}, \bold{25(4)}, 530-540. <doi:10.1080/10705511.2017.1403287>

Shun, Z., and McCullagh, P. (1995). Laplace approximation of high dimensional integrals. \emph{Journal of the Royal Statistical Society: Series B (Methodological)}, \bold{57(4)}, 749-760. <doi:10.1111/j.2517-6161.1995.tb02060.x>
}

\keyword{ package }